es6 comply
