+++ package.json
@@ -8,6 +8,7 @@
     "test:db": "cross-env NODE_ENV=test sequelize db:migrate:undo:all &&cross-env NODE_ENV=test sequelize db:migrate && npm run seed:testdb",
     "test": "npm run test:db &&cross-env NODE_ENV=test mocha  bin/test --exit",
     "start": "npm run production",
+    "start": "node_modules/.bin/webpack --config webpack.config.js",
     "start:dev": "nodemon --exec babel-node bin/dev",
     "clean": "rimraf ./dist",
     "build": "npm run clean mkdir dist && babel server -s -d dist",
@@ -42,7 +43,8 @@
     "mocha": "^4.0.1",
     "rimraf": "^2.6.2",
     "sequelize-cli": "^3.1.0",
-    "supertest": "^3.0.0"
+    "supertest": "^3.0.0",
+    "webpack": "^3.10.0"
   },
   "dependencies": {
     "bcrypt": "^1.0.3",
