+++ server/routes/index.js
@@ -1,8 +1,5 @@
-import { createEvent, modifyEvent, getEvent, deleteEvent, getAllEvents } from '../controllers/events';
-import { createCenter, modifyCenter, getAllCenters, getCenter, deleteCenter } from '../controllers/centers';
-import { createUser, signIn } from '../controllers/users';
-import { cleanData, validateToken, isAdmin, determineUser, validateUserFields, validateEventFields } from '../middlewares/';
-
+import { centerControllers, eventControllers, userControllers } from '../controllers';
+import { centerMiddlewares, eventMiddlewares, userMiddlewares, cleanData } from '../middlewares/';
 import app from './../../index';
 
 module.exports = (app) => {
@@ -14,24 +11,24 @@ module.exports = (app) => {
   });
 
   // users route
-  app.post('/api/v1/users/signup/', cleanData, validateUserFields, createUser);
-  app.post('/api/v1/users/signin/', cleanData, signIn);
+  app.post('/api/v1/users/signup/', cleanData, userMiddlewares.validateUserFields, userControllers.createUser);
+  app.post('/api/v1/users/signin/', cleanData, userControllers.signIn);
   // add admin
-  app.post('/api/v1/users/admin', cleanData, validateToken, isAdmin, validateUserFields, createUser); 
+  app.post('/api/v1/users/admin', cleanData, userMiddlewares.validateToken, userMiddlewares.isAdmin, userMiddlewares.validateUserFields, userControllers.createUser);
 
   // centers route
-  app.post('/api/v1/centers/', cleanData, validateToken, isAdmin, createCenter);
-  app.put('/api/v1/centers/:centerId', cleanData, validateToken, isAdmin, modifyCenter);
-  app.get('/api/v1/centers/:centerId', cleanData, determineUser, getCenter);
-  app.delete('/api/v1/centers/:centerId', cleanData, validateToken, isAdmin, deleteCenter);
-  app.get('/api/v1/centers/', cleanData, validateToken, getAllCenters);
+  app.post('/api/v1/centers/', cleanData, userMiddlewares.validateToken, userMiddlewares.isAdmin, centerControllers.createCenter);
+  app.put('/api/v1/centers/:centerId', cleanData, userMiddlewares.validateToken, userMiddlewares.isAdmin, centerControllers.modifyCenter);
+  app.get('/api/v1/centers/:centerId', cleanData, userMiddlewares.determineUser, centerControllers.getCenter);
+  app.delete('/api/v1/centers/:centerId', cleanData, userMiddlewares.validateToken, userMiddlewares.isAdmin, centerControllers.deleteCenter);
+  app.get('/api/v1/centers/', cleanData, userMiddlewares.validateToken, centerControllers.getAllCenters);
 
   // event route
-  app.post('/api/v1/events/', cleanData, validateToken, validateEventFields, createEvent);
-  app.put('/api/v1/events/:eventId', cleanData, validateToken, validateEventFields, modifyEvent);
-  app.get('/api/v1/events/:eventId', cleanData, validateToken, getEvent);
-  app.delete('/api/v1/events/:eventId', cleanData, validateToken, deleteEvent);
-  app.get('/api/v1/events/', cleanData, validateToken, getAllEvents);  
+  app.post('/api/v1/events/', cleanData, userMiddlewares.validateToken, eventMiddlewares.validateEventFields, eventControllers.createEvent);
+  app.put('/api/v1/events/:eventId', cleanData, userMiddlewares.validateToken, eventMiddlewares.validateEventFields, eventControllers.modifyEvent);
+  app.get('/api/v1/events/:eventId', cleanData, userMiddlewares.validateToken, eventControllers.getEvent);
+  app.delete('/api/v1/events/:eventId', cleanData, userMiddlewares.validateToken, eventControllers.deleteEvent);
+  app.get('/api/v1/events/', cleanData, userMiddlewares.validateToken, eventControllers.getAllEvents);
 
   // other routes (404)
   app.all('*', (req, res) => {
